new moving of files

+++ scripts/build/extras/pip.sh
@@ -20,10 +20,21 @@ if [ "$SCUDUM_CROSS" == "1" ]; then
     PYTHON_LIB=/usr/lib/python$PYTHON_VERSION/site-packages
     PYTHON_TOOLS_LIB=/tools/lib/python$PYTHON_VERSION/site-packages
 
-    rm -rf $PYTHON_LIB/pkg_resources* && mv $PYTHON_TOOLS_LIB/pkg_resources* $PYTHON_LIB
-    rm -rf $PYTHON_LIB/easy_install* && mv $PYTHON_TOOLS_LIB/easy_install* $PYTHON_LIB
-    rm -rf $PYTHON_LIB/setuptools* && mv $PYTHON_TOOLS_LIB/setuptools* $PYTHON_LIB
-    rm -rf $PYTHON_LIB/pip* && mv $PYTHON_TOOLS_LIB/pip* $PYTHON_LIB
+    if [ -e $PYTHON_TOOLS_LIB/pkg_resources* ]; then
+        rm -rf $PYTHON_LIB/pkg_resources* && mv $PYTHON_TOOLS_LIB/pkg_resources* $PYTHON_LIB
+    fi
+    if [ -e $PYTHON_TOOLS_LIB/easy_install* ]; then
+        rm -rf $PYTHON_LIB/easy_install* && mv $PYTHON_TOOLS_LIB/easy_install* $PYTHON_LIB
+    fi
+    if [ -e $PYTHON_TOOLS_LIB/setuptools* ]; then
+        rm -rf $PYTHON_LIB/setuptools* && mv $PYTHON_TOOLS_LIB/setuptools* $PYTHON_LIB
+    fi
+    if [ -e $PYTHON_TOOLS_LIB/_markerlib* ]; then
+        rm -rf $PYTHON_LIB/_markerlib* && mv $PYTHON_TOOLS_LIB/_markerlib* $PYTHON_LIB
+    fi
+    if [ -e $PYTHON_TOOLS_LIB/pip* ]; then
+        rm -rf $PYTHON_LIB/pip* && mv $PYTHON_TOOLS_LIB/pip* $PYTHON_LIB
+    fi
 
     mv /tools/bin/pip{,2*} /usr/bin
     mv /tools/bin/easy_install{,-2*} /usr/bin
